Disable EXT FS journaling
With FS journaling enabled, it will consume much IO for journaling.
But CI test nodes are one-time boot/running nodes for CI jobs,
journaling is not needed. So let's disable it.
And this commit contains other two changes:
1. commented out the commands which mounts ext3 fs as ext4;
2. change the configuration script in rc.local to be executed in
background. The purpose is to make XS is accessible if the scripts
got stuck.